metadata conversions
