+++ rust/numaflow-core/src/pipeline/isb/jetstream/reader.rs
@@ -7,7 +7,8 @@ use crate::Result;
 use crate::config::get_vertex_name;
 use crate::config::pipeline::isb::{BufferReaderConfig, CompressionType, ISBConfig, Stream};
 use crate::error::Error;
-use crate::message::{IntOffset, Message, MessageID, MessageType, Metadata, Offset, ReadAck};
+use crate::message::{IntOffset, Message, MessageID, MessageType, Offset, ReadAck};
+use crate::metadata::Metadata;
 use crate::metrics::{
     PIPELINE_PARTITION_NAME_LABEL, jetstream_isb_error_metrics_labels,
     jetstream_isb_metrics_labels, pipeline_metric_labels, pipeline_metrics,
@@ -22,7 +23,6 @@ use async_nats::jetstream::{
 use backoff::retry::Retry;
 use backoff::strategy::fixed;
 use bytes::Bytes;
-use chrono::Utc;
 use flate2::read::GzDecoder;
 use numaflow_throttling::RateLimiter;
 use prost::Message as ProtoMessage;
